+++ apig-20240327/alibabacloud_apig20240327/models/_get_dashboard_request.py
@@ -19,25 +19,25 @@ class GetDashboardRequest(DaraModel):
         upstream_cluster: str = None,
     ):
         # The language. Valid values:
-        # - zh: Chinese
-        # - en: English
+        # - zh: Chinese.
+        # - en: English.
         self.accept_language = accept_language
         # The unique ID of the API.
         self.api_id = api_id
         # The filter configuration.
         self.filter = filter
         # The dashboard name. Valid values:
-        # - LOG: access log
-        # - PLUGIN: plug-in log
+        # - LOG: access log.
+        # - PLUGIN: plugin log.
         self.name = name
-        # The plug-in type ID.
+        # The plugin type ID.
         self.plugin_class_id = plugin_class_id
-        # The plug-in ID.
+        # The plugin ID.
         self.plugin_id = plugin_id
         # The route ID.
         self.route_id = route_id
         # The dashboard source. Valid values:
-        # - SLS: log dashboard
+        # - SLS: log dashboard.
         self.source = source
         # The AI service identifier.
         self.upstream_cluster = upstream_cluster