+++ apig-20240327/alibabacloud_apig20240327/models/_list_tag_resources_shrink_request.py
@@ -12,13 +12,13 @@ class ListTagResourcesShrinkRequest(DaraModel):
         resource_type: str = None,
         tag_shrink: str = None,
     ):
-        # The token for the next query.
+        # The token for the next query start position.
         self.next_token = next_token
-        # The resource ID. You can specify up to 50 subkeys.
+        # The resource ID. Up to 50 subkeys are supported. At least one of ResourceId and Tag must be provided. If both are empty, the API returns InvalidParameter.BothEmpty (400).
         self.resource_id_shrink = resource_id_shrink
         # The resource type.
         self.resource_type = resource_type
-        # The list of labels to add. You can specify up to 20 subkeys.
+        # The label list. Up to 20 subkeys are supported. At least one of ResourceId and Tag must be provided. If both are empty, the API returns InvalidParameter.BothEmpty (400).
         self.tag_shrink = tag_shrink
 
     def validate(self):